changed validation level, set in sendMessage method
