+++ packages/rocketchat-api/package.js
@@ -31,8 +31,7 @@ Package.onUse(function(api) {
 	api.addFiles('server/v1/channels.js', 'server');
 	api.addFiles('server/v1/rooms.js', 'server');
 	api.addFiles('server/v1/subscriptions.js', 'server');
-	api.addFiles('server/v1/chat/chat.js', 'server');
-	api.addFiles('server/v1/chat/chat.helper.js', 'server');
+	api.addFiles('server/v1/chat.js', 'server');
 	api.addFiles('server/v1/commands.js', 'server');
 	api.addFiles('server/v1/groups.js', 'server');
 	api.addFiles('server/v1/im.js', 'server');
