+++ tests/end-to-end/api/07-incoming-integrations.js
@@ -16,6 +16,7 @@ describe('Incoming Integrations', function() {
 				type: 'webhook-incoming',
 				name: 'Incoming test',
 				enabled: true,
+				alias: 'test',
 				username: 'rocket.cat',
 				scriptEnabled: false,
 				channel: '#general'
